test: Add some initial unit tests for toxic.
These don't do much yet, but will help debug/fix unicode stuff.
